Moved code into appropriate headers (SOt instead of SO3 for now)
